Updated styling of Register and divided into two more components
Divided the Register component into a RegisterForm and a VerifyEmailPage
