+++ src/features/auth/VerifyEmailPage.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function VerifyEmailPage() {
+  return <div>Verify email!</div>;
+}
